+++ build.sh
@@ -6,6 +6,6 @@ cd $DIR
 mkdir build
 cd build
 cmake -DCMAKE_BUILD_TYPE=Release ..
-make -j48
+make -j 48
 cd $DIR
 bash compile.sh
