+++ compile.sh
@@ -6,4 +6,4 @@ cd $DIR
 mkdir -p build/release
 cd build/release
 cmake -DCMAKE_BUILD_TYPE=Release -DFORCE_TESTS=OFF ../..
-make -j48
+make -j 48
