+++ src/include/operators.h
@@ -200,7 +200,3 @@ class Checksum : public Operator {
 
         const std::vector<uint64_t> &check_sums() { return check_sums_; }
 };
-
-void reset_time();
-
-void display_time();
