+++ src/main/main.cpp
@@ -3,7 +3,7 @@
 #include "joiner.h"
 #include "parser.h"
 #include <vector>
-#include "operators.h"
+#include "utils.h"
 
 int main(int argc, char *argv[]) {
     Joiner joiner;
