+++ src/utils.cpp
@@ -2,6 +2,14 @@
 
 #include <iostream>
 
+static double filter_time = 0.0;
+static double join_prep_time = 0.0, self_join_prep_time = 0.0;
+static double join_materialization_time = 0.0, join_probing_time = 0.0, join_build_time = 0.0;
+static double self_join_materialization_time = 0.0, self_join_probing_time = 0.0;
+static double check_sum_time = 0.0;
+
+using namespace::std;
+
 // Create a dummy column
 static void createColumn(std::vector<uint64_t *> &columns,
                          uint64_t num_tuples) {
@@ -29,5 +37,85 @@ void Utils::storeRelation(std::ofstream &out, Relation &r, unsigned i) {
     r.dumpSQL(base_name, i);
     std::cout << base_name << "\n";
     out << base_name << "\n";
+}
+
+
+
+// Timer
+double * get_filter_time() {
+    return &filter_time;
+}
+
+
+double * get_self_join_prep_time() {
+    return &self_join_prep_time;
+}
+
+
+double * get_self_join_probing_time() {
+    return &self_join_probing_time;
+}
+
+
+double * get_self_join_materialization_time() {
+    return &self_join_materialization_time;
+}
+
+
+double * get_join_prep_time() {
+    return &join_prep_time;
+}
+
+
+double * get_join_probing_time() {
+    return &join_probing_time;
+}
+
+
+double * get_join_build_time() {
+    return &join_build_time;
+}
+
+
+double * get_join_materialization_time() {
+    return &join_materialization_time;
+}
+
+
+double * get_checksum_time() {
+    return &check_sum_time;
+}
+
+
+
+void reset_time() {
+    filter_time = 0.0;
+    self_join_prep_time = 0.0;
+    self_join_probing_time = 0.0;
+    self_join_materialization_time = 0.0;
+    join_prep_time = 0.0;
+    join_probing_time = 0.0;
+    join_build_time = 0.0;
+    join_materialization_time = 0.0;
+    check_sum_time = 0.0;
+}
+
+void display_time() {
+    double join_time = join_prep_time + join_probing_time + join_materialization_time + join_build_time;
+    double self_join_time = self_join_prep_time + self_join_probing_time + self_join_materialization_time;
+    double total_time = filter_time + self_join_time + join_time + check_sum_time;
+    cerr << endl;
+    cerr << "Total tracked time = " << total_time << " sec." << endl;
+    cerr << "    FilterScan time = " << filter_time << " sec." << endl;
+    cerr << "    SelfJoin time = " << self_join_time  << " sec." << endl;
+    cerr << "        Preparation time = " << self_join_prep_time << " sec." << endl;
+    cerr << "        Probing time = " << self_join_probing_time << " sec." << endl;
+    cerr << "        Materialization time = " << self_join_materialization_time << " sec." << endl;
+    cerr << "    Join time = " << join_time << " sec." << endl;
+    cerr << "        Preparation time = " << join_prep_time << " sec." << endl;
+    cerr << "        Building time = " << join_build_time << " sec." << endl;
+    cerr << "        Probing time = " << join_probing_time << " sec." << endl;
+    cerr << "        Materialization time = " << join_materialization_time << " sec." << endl;
+    cerr << "    Checksum time = " << check_sum_time << " sec." << endl;
 }
